ml-service/requirements.txt
flask
flask-cors
joblib
pandas
numpy
scikit-learn
xgboost
python-dotenv
